# api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Heimdallr API
  version: 1.0.0
  description: API for managing applications, releases, compliance reports, servers, agents, providers, automations, and automation jobs.
servers:
  - url: /api
    description: API base path
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/analytics/automation:
    get:
      tags:
        - Analytics
      summary: Get automation analytics overview
      operationId: getAutomationAnalyticsOverview
      responses:
        "200":
          description: Automation analytics overview returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsAutomationDataResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/analytics/automation/{automation_id}:
    get:
      tags:
        - Analytics
      summary: Get analytics overview for one automation
      operationId: getAutomationAnalyticsOverviewByID
      parameters:
        - $ref: "#/components/parameters/AutomationIDPath"
      responses:
        "200":
          description: Automation analytics returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsAutomationDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/analytics/compliance:
    get:
      tags:
        - Analytics
      summary: Get compliance analytics overview
      operationId: getComplianceAnalyticsOverview
      responses:
        "200":
          description: Compliance analytics overview returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceAnalyticsDataResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API token passed as a Bearer token in the Authorization header
    SessionCookie:
      type: apiKey
      in: cookie
      name: heimdallr_session
      description: HttpOnly browser session cookie
  parameters:
...
    AutomationIDPath:
      name: automation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: invalid request body
...
    LocationJobAnalytics:
      type: object
      required:
        - location
        - total_jobs
        - successful_jobs
        - failed_jobs
        - started_jobs
        - success_rate
      properties:
        location:
          type: string
        total_jobs:
          type: integer
          minimum: 0
        successful_jobs:
          type: integer
          minimum: 0
        failed_jobs:
          type: integer
          minimum: 0
        started_jobs:
          type: integer
          minimum: 0
        success_rate:
          type: number
          format: double
    AutomationJobAnalytics:
      type: object
      required:
        - automation_id
        - automation
        - provider
        - total_jobs
        - successful_jobs
        - failed_jobs
        - started_jobs
        - success_rate
      properties:
        automation_id:
          type: string
        automation:
          type: string
        provider:
          type: string
        total_jobs:
          type: integer
          minimum: 0
        successful_jobs:
          type: integer
          minimum: 0
        failed_jobs:
          type: integer
          minimum: 0
        started_jobs:
          type: integer
          minimum: 0
        success_rate:
          type: number
          format: double
    AutomationAnalytics:
      type: object
      required:
        - total_automations
        - total_jobs
        - successful_jobs
        - failed_jobs
        - started_jobs
        - success_rate
        - by_location
        - by_automation
      properties:
        total_automations:
          type: integer
          minimum: 0
        total_jobs:
          type: integer
          minimum: 0
        successful_jobs:
          type: integer
          minimum: 0
        failed_jobs:
          type: integer
          minimum: 0
        started_jobs:
          type: integer
          minimum: 0
        success_rate:
          type: number
          format: double
        by_location:
          type: array
          items:
            $ref: "#/components/schemas/LocationJobAnalytics"
        by_automation:
          type: array
          items:
            $ref: "#/components/schemas/AutomationJobAnalytics"
    AnalyticsAutomationDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/AutomationAnalytics"
...
    ComplianceAnalytics:
      type: object
      required:
        - total_applications
        - total_releases
        - total_reports
        - successful_reports
        - failed_reports
        - started_reports
        - success_rate
        - by_application
      properties:
        total_applications:
          type: integer
          minimum: 0
        total_releases:
          type: integer
          minimum: 0
        total_reports:
          type: integer
          minimum: 0
        successful_reports:
          type: integer
          minimum: 0
        failed_reports:
          type: integer
          minimum: 0
        started_reports:
          type: integer
          minimum: 0
        success_rate:
          type: number
          format: double
        by_application:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationComplianceStats"
    ApplicationComplianceStats:
      type: object
      required:
        - application_id
        - application
        - latest_release_id
        - latest_version
        - total_reports
        - successful_reports
        - failed_reports
        - started_reports
        - success_rate
      properties:
        application_id:
          type: string
        application:
          type: string
        latest_release_id:
          type: string
        latest_version:
          type: string
        total_reports:
          type: integer
          minimum: 0
        successful_reports:
          type: integer
          minimum: 0
        failed_reports:
          type: integer
          minimum: 0
        started_reports:
          type: integer
          minimum: 0
        success_rate:
          type: number
          format: double
    ComplianceAnalyticsDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/ComplianceAnalytics"
...