# internal/analytics/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes    bearerAuthContextKey    = "BearerAuth.Scopes"
	SessionCookieScopes sessionCookieContextKey = "SessionCookie.Scopes"
)

// AnalyticsAutomationDataResponse defines model for AnalyticsAutomationDataResponse.
type AnalyticsAutomationDataResponse struct {
	Data AutomationAnalytics `json:"data"`
}

// ApplicationComplianceStats defines model for ApplicationComplianceStats.
type ApplicationComplianceStats struct {
	Application       string  `json:"application"`
	ApplicationId     string  `json:"application_id"`
	FailedReports     int     `json:"failed_reports"`
	LatestReleaseId   string  `json:"latest_release_id"`
	LatestVersion     string  `json:"latest_version"`
	StartedReports    int     `json:"started_reports"`
	SuccessRate       float64 `json:"success_rate"`
	SuccessfulReports int     `json:"successful_reports"`
	TotalReports      int     `json:"total_reports"`
}

// AutomationAnalytics defines model for AutomationAnalytics.
type AutomationAnalytics struct {
	ByAutomation     []AutomationJobAnalytics `json:"by_automation"`
	ByLocation       []LocationJobAnalytics   `json:"by_location"`
	FailedJobs       int                      `json:"failed_jobs"`
	StartedJobs      int                      `json:"started_jobs"`
	SuccessRate      float64                  `json:"success_rate"`
	SuccessfulJobs   int                      `json:"successful_jobs"`
	TotalAutomations int                      `json:"total_automations"`
	TotalJobs        int                      `json:"total_jobs"`
}

// AutomationJobAnalytics defines model for AutomationJobAnalytics.
type AutomationJobAnalytics struct {
	Automation     string  `json:"automation"`
	AutomationId   string  `json:"automation_id"`
	FailedJobs     int     `json:"failed_jobs"`
	Provider       string  `json:"provider"`
	StartedJobs    int     `json:"started_jobs"`
	SuccessRate    float64 `json:"success_rate"`
	SuccessfulJobs int     `json:"successful_jobs"`
	TotalJobs      int     `json:"total_jobs"`
}

// ComplianceAnalytics defines model for ComplianceAnalytics.
type ComplianceAnalytics struct {
	ByApplication     []ApplicationComplianceStats `json:"by_application"`
	FailedReports     int                          `json:"failed_reports"`
	StartedReports    int                          `json:"started_reports"`
	SuccessRate       float64                      `json:"success_rate"`
	SuccessfulReports int                          `json:"successful_reports"`
	TotalApplications int                          `json:"total_applications"`
	TotalReleases     int                          `json:"total_releases"`
	TotalReports      int                          `json:"total_reports"`
}

// ComplianceAnalyticsDataResponse defines model for ComplianceAnalyticsDataResponse.
type ComplianceAnalyticsDataResponse struct {
	Data ComplianceAnalytics `json:"data"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// LocationJobAnalytics defines model for LocationJobAnalytics.
type LocationJobAnalytics struct {
	FailedJobs     int     `json:"failed_jobs"`
	Location       string  `json:"location"`
	StartedJobs    int     `json:"started_jobs"`
	SuccessRate    float64 `json:"success_rate"`
	SuccessfulJobs int     `json:"successful_jobs"`
	TotalJobs      int     `json:"total_jobs"`
}

// AutomationIDPath defines model for AutomationIDPath.
type AutomationIDPath = openapi_types.UUID

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// InternalServerError defines model for InternalServerError.
type InternalServerError = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// Unauthorized defines model for Unauthorized.
type Unauthorized = ErrorResponse

// bearerAuthContextKey is the context key for BearerAuth security scheme
type bearerAuthContextKey string

// sessionCookieContextKey is the context key for SessionCookie security scheme
type sessionCookieContextKey string

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get automation analytics overview
	// (GET /v1/analytics/automation)
	GetAutomationAnalyticsOverview(c *gin.Context)
	// Get analytics overview for one automation
	// (GET /v1/analytics/automation/{automation_id})
	GetAutomationAnalyticsOverviewByID(c *gin.Context, automationId AutomationIDPath)
	// Get compliance analytics overview
	// (GET /v1/analytics/compliance)
	GetComplianceAnalyticsOverview(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetAutomationAnalyticsOverview operation middleware
func (siw *ServerInterfaceWrapper) GetAutomationAnalyticsOverview(c *gin.Context) {

	c.Set(string(BearerAuthScopes), []string{})

	c.Set(string(SessionCookieScopes), []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetAutomationAnalyticsOverview(c)
}

// GetAutomationAnalyticsOverviewByID operation middleware
func (siw *ServerInterfaceWrapper) GetAutomationAnalyticsOverviewByID(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "automation_id" -------------
	var automationId AutomationIDPath

	err = runtime.BindStyledParameterWithOptions("simple", "automation_id", c.Param("automation_id"), &automationId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter automation_id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(string(BearerAuthScopes), []string{})

	c.Set(string(SessionCookieScopes), []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetAutomationAnalyticsOverviewByID(c, automationId)
}

// GetComplianceAnalyticsOverview operation middleware
func (siw *ServerInterfaceWrapper) GetComplianceAnalyticsOverview(c *gin.Context) {

	c.Set(string(BearerAuthScopes), []string{})

	c.Set(string(SessionCookieScopes), []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetComplianceAnalyticsOverview(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/v1/analytics/automation", wrapper.GetAutomationAnalyticsOverview)
	router.GET(options.BaseURL+"/v1/analytics/automation/:automation_id", wrapper.GetAutomationAnalyticsOverviewByID)
	router.GET(options.BaseURL+"/v1/analytics/compliance", wrapper.GetComplianceAnalyticsOverview)
}

type BadRequestJSONResponse ErrorResponse

type InternalServerErrorJSONResponse ErrorResponse

type NotFoundJSONResponse ErrorResponse

type UnauthorizedJSONResponse ErrorResponse

type GetAutomationAnalyticsOverviewRequestObject struct {
}

type GetAutomationAnalyticsOverviewResponseObject interface {
	VisitGetAutomationAnalyticsOverviewResponse(w http.ResponseWriter) error
}

type GetAutomationAnalyticsOverview200JSONResponse AnalyticsAutomationDataResponse

func (response GetAutomationAnalyticsOverview200JSONResponse) VisitGetAutomationAnalyticsOverviewResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetAutomationAnalyticsOverview500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response GetAutomationAnalyticsOverview500JSONResponse) VisitGetAutomationAnalyticsOverviewResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetAutomationAnalyticsOverviewByIDRequestObject struct {
	AutomationId AutomationIDPath `json:"automation_id"`
}

type GetAutomationAnalyticsOverviewByIDResponseObject interface {
	VisitGetAutomationAnalyticsOverviewByIDResponse(w http.ResponseWriter) error
}

type GetAutomationAnalyticsOverviewByID200JSONResponse AnalyticsAutomationDataResponse

func (response GetAutomationAnalyticsOverviewByID200JSONResponse) VisitGetAutomationAnalyticsOverviewByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetAutomationAnalyticsOverviewByID400JSONResponse struct{ BadRequestJSONResponse }

func (response GetAutomationAnalyticsOverviewByID400JSONResponse) VisitGetAutomationAnalyticsOverviewByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetAutomationAnalyticsOverviewByID404JSONResponse struct{ NotFoundJSONResponse }

func (response GetAutomationAnalyticsOverviewByID404JSONResponse) VisitGetAutomationAnalyticsOverviewByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetAutomationAnalyticsOverviewByID500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response GetAutomationAnalyticsOverviewByID500JSONResponse) VisitGetAutomationAnalyticsOverviewByIDResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetComplianceAnalyticsOverviewRequestObject struct {
}

type GetComplianceAnalyticsOverviewResponseObject interface {
	VisitGetComplianceAnalyticsOverviewResponse(w http.ResponseWriter) error
}

type GetComplianceAnalyticsOverview200JSONResponse ComplianceAnalyticsDataResponse

func (response GetComplianceAnalyticsOverview200JSONResponse) VisitGetComplianceAnalyticsOverviewResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetComplianceAnalyticsOverview401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetComplianceAnalyticsOverview401JSONResponse) VisitGetComplianceAnalyticsOverviewResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type GetComplianceAnalyticsOverview500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response GetComplianceAnalyticsOverview500JSONResponse) VisitGetComplianceAnalyticsOverviewResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Get automation analytics overview
	// (GET /v1/analytics/automation)
	GetAutomationAnalyticsOverview(ctx context.Context, request GetAutomationAnalyticsOverviewRequestObject) (GetAutomationAnalyticsOverviewResponseObject, error)
	// Get analytics overview for one automation
	// (GET /v1/analytics/automation/{automation_id})
	GetAutomationAnalyticsOverviewByID(ctx context.Context, request GetAutomationAnalyticsOverviewByIDRequestObject) (GetAutomationAnalyticsOverviewByIDResponseObject, error)
	// Get compliance analytics overview
	// (GET /v1/analytics/compliance)
	GetComplianceAnalyticsOverview(ctx context.Context, request GetComplianceAnalyticsOverviewRequestObject) (GetComplianceAnalyticsOverviewResponseObject, error)
}

type StrictHandlerFunc func(ctx *gin.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictGinServerOptions struct {
	// RequestErrorHandlerFunc is called when a request cannot be parsed or
	// decoded. It is invoked for JSON bind failures, form parse/bind errors,
	// multipart reader errors, media type parse errors, missing multipart
	// boundaries, and request body read errors. The default returns 400.
	RequestErrorHandlerFunc func(ctx *gin.Context, err error)
	// HandlerErrorFunc is called when the application handler (or any
	// middleware wrapping it) returns a non-nil error. The default returns 500.
	HandlerErrorFunc func(ctx *gin.Context, err error)
	// ResponseErrorHandlerFunc is called when the response object fails to
	// serialize (Visit*Response returns an error) or when the handler returns
	// an unexpected response type. The default returns 500.
	ResponseErrorHandlerFunc func(ctx *gin.Context, err error)
}

func defaultStrictGinRequestErrorHandler(ctx *gin.Context, err error) {
	if _, ok := err.(*http.MaxBytesError); ok {
		ctx.JSON(http.StatusRequestEntityTooLarge, gin.H{"error": "request body too large"})
		return
	}
	ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictGinServerOptions{
		RequestErrorHandlerFunc: defaultStrictGinRequestErrorHandler,
		HandlerErrorFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"error": http.StatusText(http.StatusInternalServerError)})
		},
		ResponseErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"error": http.StatusText(http.StatusInternalServerError)})
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictGinServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = defaultStrictGinRequestErrorHandler
	}
	if options.HandlerErrorFunc == nil {
		options.HandlerErrorFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"error": http.StatusText(http.StatusInternalServerError)})
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"error": http.StatusText(http.StatusInternalServerError)})
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictGinServerOptions
}

// GetAutomationAnalyticsOverview operation middleware
func (sh *strictHandler) GetAutomationAnalyticsOverview(ctx *gin.Context) {
	var request GetAutomationAnalyticsOverviewRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetAutomationAnalyticsOverview(ctx.Request.Context(), request.(GetAutomationAnalyticsOverviewRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetAutomationAnalyticsOverview")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetAutomationAnalyticsOverviewResponseObject); ok {
		if err := validResponse.VisitGetAutomationAnalyticsOverviewResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetAutomationAnalyticsOverviewByID operation middleware
func (sh *strictHandler) GetAutomationAnalyticsOverviewByID(ctx *gin.Context, automationId AutomationIDPath) {
	var request GetAutomationAnalyticsOverviewByIDRequestObject

	request.AutomationId = automationId

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetAutomationAnalyticsOverviewByID(ctx.Request.Context(), request.(GetAutomationAnalyticsOverviewByIDRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetAutomationAnalyticsOverviewByID")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetAutomationAnalyticsOverviewByIDResponseObject); ok {
		if err := validResponse.VisitGetAutomationAnalyticsOverviewByIDResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetComplianceAnalyticsOverview operation middleware
func (sh *strictHandler) GetComplianceAnalyticsOverview(ctx *gin.Context) {
	var request GetComplianceAnalyticsOverviewRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetComplianceAnalyticsOverview(ctx.Request.Context(), request.(GetComplianceAnalyticsOverviewRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetComplianceAnalyticsOverview")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetComplianceAnalyticsOverviewResponseObject); ok {
		if err := validResponse.VisitGetComplianceAnalyticsOverviewResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}